backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pymongo
pydantic